server/requirements.txt
bcrypt==5.0.0
email-validator==2.3.0
fastapi==0.116.1
passlib==1.7.4
pydantic==2.11.7
pydantic-settings==2.13.1
python-jose==3.5.0
python-multipart==0.0.20
SQLAlchemy==2.0.48
uvicorn==0.35.0
pytest==8.3.3
httpx==0.27.2